% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaPvals.R
\name{anovaPvals}
\alias{anovaPvals}
\title{Extract P-values for a Specified Factor from Anova Results}
\usage{
anovaPvals(anova_results, factor_name)
}
\arguments{
\item{anova_results}{A list of Anova results, typically obtained from the `car::Anova()` function.}

\item{factor_name}{A character string specifying the name of the factor to extract p-values for.}
}
\value{
A list of numeric p-values for the specified factor from each Anova result in the input list.
}
\description{
This function extracts p-values for a specific factor from a list of Anova results.
It returns a list of numeric p-values for the specified factor, excluding any NA values.
}
\examples{
\dontrun{
data <- data.frame(dyad_id = rep(1:10, each = 2),
                   x = rnorm(20),
                   y = rnorm(20))
results <- randBoot(data, "dyad_id", y ~ x, model_type = "lm",
                    focal_cols = "x", opposite_cols = "y")
pvalues <- anovaPvals(results$anova_results, "opposite_group")
}
}
