% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{averages}
\alias{averages}
\title{Calculate Averages and Standard Errors from Bootstrap Results}
\usage{
averages(bootstrap_results, max_iterations = 10000)
}
\arguments{
\item{bootstrap_results}{A list of model summaries obtained from bootstrapping.}

\item{max_iterations}{An integer specifying the maximum number of iterations allowed in the function. This helps in preventing infinite recursion. Default is set to 10000.}
}
\value{
A list containing two matrices: 'averages' with the average values of the coefficients and 'standard_errors' with the standard errors of the coefficients.
}
\description{
This function calculates the averages and standard errors of coefficients from the results of bootstrapped models.
}
\examples{
\dontrun{
data <- data.frame(y = rnorm(100), x = rnorm(100))
results <- bootstrap_lm(data, y ~ x, 100)
avg_se <- averages(results)
}
}
