% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histMeans.R
\name{histMeans}
\alias{histMeans}
\title{Plot Histogram for a Specific Level}
\usage{
histMeans(all_means, num_var, cat_var, level)
}
\arguments{
\item{all_means}{A data frame containing means for each dataset and level (typically the \code{all_means}
output from the \code{repData} function).}

\item{num_var}{The name of the numerical variable for which the means should be plotted.}

\item{cat_var}{The name of the categorical variable which determines the different levels.}

\item{level}{The specific level of the categorical variable for which the histogram should be plotted.}
}
\value{
A ggplot object representing the histogram.
}
\description{
This function plots a histogram for the means of a specified numerical variable for a specific level of
a specified categorical variable.
}
\examples{
\dontrun{
histMeans(result$all_means, "focal_dant", "focal_morph", "Fw")
}

}
