% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histPvals.R
\name{histPvals}
\alias{histPvals}
\title{Plot Histogram of p-values for a Specified Coefficient from Bootstrap Results}
\usage{
histPvals(bootstrap_results, coeff_name, max_iterations = 10000)
}
\arguments{
\item{bootstrap_results}{A list of model summaries obtained from bootstrapping.}

\item{coeff_name}{A character string specifying the name of the coefficient for which the histogram of p-values is to be plotted.}

\item{max_iterations}{An integer specifying the maximum number of iterations allowed in the function. This helps in preventing infinite recursion. Default is set to 10000.}
}
\value{
A histogram plot showing the distribution of p-values for the specified coefficient.
}
\description{
This function extracts the p-values for a specified coefficient from bootstrapped model results and plots a histogram.
}
\examples{
\dontrun{
data <- data.frame(y = rnorm(100), x = rnorm(100))
results <- bootstrap_lm(data, y ~ x, 100)
histPvals(results, "x")
}
}
