% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanOfMeans.R
\name{MeanOfMeans}
\alias{MeanOfMeans}
\title{MeanOfMeans function}
\usage{
MeanOfMeans(all_means_df, num_var, cat_var1, cat_var2 = NULL)
}
\arguments{
\item{all_means_df}{A data frame containing the mean values of the numerical variable for each level (or combination of levels) of the categorical variable(s) across multiple datasets.}

\item{num_var}{The name of the numerical variable as a string.}

\item{cat_var1}{The name of the first categorical variable as a string.}

\item{cat_var2}{The name of the second categorical variable as a string (optional).}
}
\value{
A data frame containing the overall mean value(s) of the numerical variable for each level (or combination of levels) of the categorical variable(s) across all datasets.
}
\description{
This function takes a data frame containing the mean values of a numerical variable for each level (or combination of levels)
of one or two categorical variables across multiple datasets, and computes the overall mean value(s) across all datasets.
}
\examples{
\dontrun{
MeanOfMeans(each_dataset_means, "dant", "Group")
MeanOfMeans(each_dataset_means, "dant", "Group", "Song")
}

}
