% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Randomly Assign Roles within Dyads}
\usage{
rand(data, dyad_id_col, max_iterations = 10000)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{dyad_id_col}{A character string specifying the name of the column that contains dyad identifiers.}

\item{max_iterations}{An integer specifying the maximum number of times the function can be called. Defaults to 10,000.}
}
\value{
A dataframe with the original columns prefixed by 'focal_' or 'opposite_' depending on the role assigned.
Columns from the same row in the original dataframe will be in separate rows in the output,
with their roles indicated by the prefix.
}
\description{
This function takes a dataset and randomly assigns roles within dyads, returning
a dataset where each dyad has a 'focal' and an 'opposite' member.
}
\examples{
\dontrun{
dyad_data <- data.frame(dyad_id = c(1,1,2,2), var1 = c("A", "B", "C", "D"))
new_data <- rand(dyad_data, dyad_id_col = "dyad_id")
}

}
