% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randBoot.R
\name{randBoot}
\alias{randBoot}
\title{Randomized Bootstrap Function}
\usage{
randBoot(
  data,
  dyad_id_col,
  model_formula,
  model_type = "lm",
  family = NULL,
  n_bootstraps = 1000,
  focal_cols,
  opposite_cols,
  RE_formula = NULL,
  control = NULL,
  max_iterations = 10000,
  max_time_seconds = 3600
)
}
\arguments{
\item{data}{A data frame containing the dyadic data.}

\item{dyad_id_col}{A character string specifying the column name which identifies the dyad.}

\item{model_formula}{A formula specifying the model to be run.}

\item{model_type}{A character string specifying the type of model. One of "lm", "glm", "lmer", or "glmer".}

\item{family}{A family object passed to glm or glmer, if applicable.}

\item{n_bootstraps}{An integer specifying the number of bootstraps to run. Default is 1000.}

\item{focal_cols}{A character vector of column names for the focal individual.}

\item{opposite_cols}{A character vector of column names for the opposite individual.}

\item{RE_formula}{A formula specifying the random effects, if using a mixed-effects model.}

\item{control}{Control arguments passed to lmer or glmer.}

\item{max_iterations}{An integer specifying the maximum number of iterations. Default is 10000.}

\item{max_time_seconds}{An integer specifying the maximum execution time in seconds. Default is 3600.}
}
\value{
A list containing two lists:
\itemize{
  \item bootstrap_results: A list of models and their summaries for each bootstrap iteration.
  \item anova_results: A list of ANOVA results for each model.
}
}
\description{
This function performs a randomized bootstrap analysis on dyadic data.
It returns both bootstrap model results and ANOVA results.
}
\examples{
\dontrun{
data <- data.frame(dyad_id = rep(1:10, each = 2),
                   x = rnorm(20),
                   y = rnorm(20))
results <- randBoot(data, "dyad_id", y ~ x, model_type = "lm",
                    focal_cols = "x", opposite_cols = "y")
}
}
