% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randMult.R
\name{randMult}
\alias{randMult}
\title{Multiple Random Dyad Assignments}
\usage{
randMult(data, dyad_id_col, num_iterations = 1, max_iterations = 10000)
}
\arguments{
\item{data}{A data frame containing the original dataset.}

\item{dyad_id_col}{The name of the column in the data that contains the dyad identifiers.}

\item{num_iterations}{An integer specifying the number of iterations for the random assignment. Default is 1.}

\item{max_iterations}{An integer specifying the maximum number of allowed iterations. Default is 10000.}
}
\value{
A list containing:
\itemize{
  \item \code{results_list}: A list of data frames, each representing an iteration of the random assignments.
}
}
\description{
This function takes a dataset and, based on dyad identifiers, randomly assigns roles of "focal"
and "opposite" to each dyad. It repeats this process for a specified number of iterations.
}
\examples{
\dontrun{
result <- randMult(my_data, "dyad_column")
}

}
