% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDataSet.R
\name{repDataSet}
\alias{repDataSet}
\title{Compute Means and Identify Closest Dataset}
\usage{
repDataSet(resultsMult, num_var, cat_var)
}
\arguments{
\item{resultsMult}{A list containing multiple datasets as a result of multiple iterations.}

\item{num_var}{The name of the numerical variable for which means are to be calculated.}

\item{cat_var}{The name of the categorical variable that defines the levels across which means are to be computed.}
}
\value{
A list containing:
\itemize{
  \item \code{each_dataset_means} - A dataframe containing the means for each iteration and level of the categorical variable.
  \item \code{closest_dataset} - The dataset from the list of input datasets that has values closest to the calculated means.
}
}
\description{
This function computes the means for a numerical variable across different levels of a categorical variable.
Additionally, it identifies the dataset that has values closest to the calculated means.
}
\examples{
# Assuming you have a list `results` containing multiple datasets
# and you want to compute means for the variable 'value' across levels of 'category':
# result_data <- repDataSet(results, "value", "category")
}
