% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanOfMeans.R
\name{MeanOfMeans}
\alias{MeanOfMeans}
\title{Calculate Overall Means for Each Level}
\usage{
MeanOfMeans(all_means_df)
}
\arguments{
\item{all_means_df}{A dataframe containing the mean values for each level of the categorical variable
for each dataset. This dataframe is typically obtained from the \code{repDataSet} function.}
}
\value{
A named numeric vector where names are the levels of the categorical variable and
the values are the overall means for each level.
}
\description{
This function computes the overall mean for each level of the categorical variable
using the provided dataframe that contains the means for each dataset.
}
\examples{
# Assuming you have the 'all_means_df' dataframe from the 'repDataSet' function:
# overall_means_vector <- MeanOfMeans(all_means_df)
}
