% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoot.R
\name{plotBoot}
\alias{plotBoot}
\title{Raincloud Plot with Optional P-value Histogram}
\usage{
plotBoot(
  closest_data,
  num_var,
  cat_var,
  overall_means,
  xlab_name = cat_var,
  ylab_name = num_var,
  main_title = NULL,
  p_values = NULL
)
}
\arguments{
\item{closest_data}{A dataframe containing the data to be plotted.}

\item{num_var}{The name of the numeric variable in the `closest_data` dataframe.}

\item{cat_var}{The name of the categorical variable in the `closest_data` dataframe.}

\item{overall_means}{A named vector of overall means for each level of the categorical variable.}

\item{xlab_name}{Optional. Name of the x-axis. Defaults to the name of `cat_var`.}

\item{ylab_name}{Optional. Name of the y-axis. Defaults to the name of `num_var`.}

\item{main_title}{Optional. Main title for the raincloud plot.}

\item{p_values}{Optional. A list or vector of p-values. If provided, a histogram of the p-values
is generated alongside the raincloud plot.}
}
\value{
A ggplot2 plot object. If `p_values` are provided, the returned plot will include
both the raincloud plot and the p-value histogram side by side.
}
\description{
This function generates a raincloud plot of the provided data. If p-values are also
provided, it generates an additional histogram of the p-values alongside the raincloud plot.
}
\examples{
\dontrun{
plotBoot(df, "numeric_var", "categorical_var", overall_means = means_vec, p_values = pvals)
}

}
