% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randboot.R
\name{randboot}
\alias{randboot}
\title{Bootstrap Random Dyad Data and Model}
\usage{
randboot(
  data,
  dyad_id_col,
  model_formula,
  model_type = "lm",
  family = NULL,
  n_bootstraps = 1000,
  focal_cols,
  opposite_cols,
  RE_formula = NULL,
  control = NULL,
  max_iterations = 10000,
  max_time_seconds = 3600
)
}
\arguments{
\item{data}{A data frame containing the dyad data.}

\item{dyad_id_col}{The column name in the data frame that identifies dyads.}

\item{model_formula}{The formula for the model to be fit.}

\item{model_type}{The type of model to be fit. This can be one of 'lm', 'glm', 'lmer', or 'glmer'.}

\item{family}{For 'glm' and 'glmer' models, the family object describing the error distribution and link function.}

\item{n_bootstraps}{The number of bootstrap iterations.}

\item{focal_cols}{The columns in the data that pertain to the focal member of the dyad.}

\item{opposite_cols}{The columns in the data that pertain to the opposite member of the dyad.}

\item{RE_formula}{The random effects formula for 'lmer' and 'glmer' models.}

\item{control}{A list of control parameters for 'lmer' and 'glmer' models.}

\item{max_iterations}{The maximum number of iterations allowed for the bootstrap process (default is 10,000).}

\item{max_time_seconds}{The maximum execution time in seconds allowed for the bootstrap process (default is 3,600 seconds or 1 hour).}
}
\value{
A list containing the bootstrapped model results for each iteration.
}
\description{
This function bootstraps dyad data by randomly assigning roles to the members
of each dyad and then fits the specified model to the bootstrapped data.
}
\examples{
\dontrun{
data <- data.frame(dyad_id = rep(1:50, each = 2), x = rnorm(100), y = rnorm(100))
result <- randboot(data, dyad_id_col = "dyad_id", model_formula = y ~ x, model_type = "lm",
                         focal_cols = "x", opposite_cols = "y")
}

}
