% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDataSet.R
\name{repDataSet}
\alias{repDataSet}
\title{repDataSet function}
\usage{
repDataSet(resultsMult, num_var, cat_var1, cat_var2 = NULL)
}
\arguments{
\item{resultsMult}{A list containing multiple datasets.}

\item{num_var}{The name of the numerical variable as a string.}

\item{cat_var1}{The name of the first categorical variable as a string.}

\item{cat_var2}{The name of the second categorical variable as a string (optional).}
}
\value{
A list containing two elements:
\itemize{
  \item each_dataset_means: A data frame containing the iteration number and the mean of the numerical variable for each level (or combination of levels) of the categorical variable(s) in each dataset.
  \item closest_dataset: The dataset from the input list that has the mean value(s) closest to the overall mean value(s) across all datasets.
}
}
\description{
This function takes a list of datasets along with names of a numerical variable and one or two categorical variables.
It computes the mean of the numerical variable for each level (or combination of levels) of the categorical variable(s)
in each dataset. It also identifies the dataset whose mean value(s) are closest to the overall mean value(s) across
all datasets.
}
\examples{
\dontrun{
repDataSet(resultsMult, "dant", "Group")
repDataSet(resultsMult, "dant", "Group", "Song")
}

}
